requirements.txt
cryptography>=41.0.0
pyOpenSSL>=23.0.0
asyncio>=3.4.3
uvloop>=0.17.0
aiohttp>=3.8.0
pyinstaller>=5.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
statistics>=1.0.3.5
python-dateutil>=2.8.2 